refactor: #10 and #11, 게시판 및 댓글 BaseResponse 추가

+++ src/main/java/doubleni/mealrecipe/service/BoardService.java
@@ -2,27 +2,19 @@ package doubleni.mealrecipe.service;
 
 import doubleni.mealrecipe.config.MD5Generator;
 import doubleni.mealrecipe.config.exception.BaseException;
-import doubleni.mealrecipe.config.exception.BaseResponse;
-import doubleni.mealrecipe.config.exception.BaseResponseStatus;
 import doubleni.mealrecipe.model.Board;
 import doubleni.mealrecipe.model.DTO.*;
 import doubleni.mealrecipe.model.User;
 import doubleni.mealrecipe.repository.BoardRepository;
-import doubleni.mealrecipe.repository.CommentRepository;
 import doubleni.mealrecipe.repository.FileRepository;
 import doubleni.mealrecipe.repository.UserRepository;
-import doubleni.mealrecipe.utils.JwtService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.IOException;
-import java.security.NoSuchAlgorithmException;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -44,23 +36,14 @@ public class BoardService {
     private final UserRepository userRepository;
     private final BoardRepository boardRepository;
     private final FileRepository fileRepository;
+    private final FileService fileService;
 
     /* TEST */
 
     // ==========================================================
 
-//    public Page<Board> findAll(PageRequest pageRequest) {
-//        return boardRepository.findAll(pageRequest);
-//    }
-//
-//    public Page<Board> findByTitle(PageRequest pageRequest) {
-//        return boardRepository.findByTitle("1", pageRequest);
-//    }
-
-
     /* 모든 게시글(리스트) 조회 */
-    @Transactional(readOnly = true)
-    public List<BoardRes> getBoards() {
+    public List<BoardRes> getBoards() throws BaseException {
         List<BoardRes> boards =
                 boardRepository.findAll()
                         .stream()
@@ -68,27 +51,48 @@ public class BoardService {
                         .collect(Collectors.toList());
 
         if (boards.isEmpty()) {
-            throw new IllegalStateException();
+            throw new BaseException(BOARD_NOT_EXISTS);
         }
         return boards;
     }
 
-    /* board id로 게시글 조회 */
-    public Board getBoardById(Long boardId) { // 내 글 불러 오는 겨
-        return boardRepository.findByBoardId(boardId)
-                .orElseThrow(() -> new IllegalArgumentException("not found : " + boardId));
+    /* 나의 게시글 조회 */
+    public List<BoardRes> getMyBoards(Long idx) throws BaseException {
+        try {
+            Optional<User> userOptional = userRepository.findById(idx);
+
+            if (userOptional.isPresent()) {
+                User user = userOptional.get();
+                List<BoardRes> myBoards =
+                        boardRepository.findBoardByUser(user)
+                                .stream()
+                                .map(BoardRes::new)
+                                .toList();
+                return myBoards;
+            } else {
+                    // 사용자를 찾지 못한 경우 에러 처리
+                    throw new BaseException(USERS_NOT_EXISTS);
+                }
+        } catch (Exception exception) {
+            throw new BaseException(BOARD_NOT_EXISTS);
+        }
+
+
     }
-    public Board getBoardByBoardId(Long boardId) throws BaseException { // 조회수 증가
+
+    /* boardId로 게시글 조회 */
+    public Board getBoardByBoardId(Long boardId, Long idx) throws BaseException {
         Optional<Board> findBoard = this.boardRepository.findByBoardId(boardId);
         if (findBoard.isPresent()) {
             Board board = findBoard.get();
-            board.setHits(board.getHits()+1);
+            if (!idx.equals(findBoard.get().getUser().getId())) {
+                board.setHits(board.getHits()+1); // 조회수 증가
+            }
             this.boardRepository.save(board);
             return board;
         } else {
             throw new BaseException(DATABASE_ERROR);
         }
-
     }
 
     /* 게시글 정렬 */
@@ -130,9 +134,6 @@ public class BoardService {
         }
         return getBoards;
     }
-
-
-
 
     /* 게시글 검색 */
     public List<BoardRes> searchBoardByUserNickname(String writer) {
@@ -177,42 +178,81 @@ public class BoardService {
 
 
     /* 게시글 추가 */
-    @Transactional
-    public void save(BoardReq addBoard, Long id) throws BaseException {
+    public BoardRes add(BoardReq addBoard, MultipartFile file, Long idx) throws BaseException {
         try {
-            Optional<User> user = userRepository.findById(id);
-            if (user.isPresent()) {
-                boardRepository.save(addBoard.toEntity(user.get()));
-                ResponseEntity.ok("-> 게시글 등록 완료");
+            Optional<User> optUser = userRepository.findById(idx);
+
+            if (optUser.isPresent()) {
+                User user = optUser.get();
+                Board board = addBoard.toEntity(user);
+
+                if (file != null) {
+                String origFilename = file.getOriginalFilename();
+                String filename = new MD5Generator(origFilename).toString();
+                String savePath = System.getProperty("user.dir") + "\\files";
+                if (!new File(savePath).exists()) {
+                    try {
+                        new File(savePath).mkdir();
+                    } catch(Exception e){
+                        e.getStackTrace();
+                    }
+                }
+                String filePath = savePath + "\\" + filename;
+                file.transferTo(new File(filePath));
+
+                FileReq fileDto = new FileReq();
+                fileDto.setOrigFilename(origFilename);
+                fileDto.setFilename(filename);
+                fileDto.setFilePath(filePath);
+
+                Long fileId = fileService.saveFile(fileDto);
+                board.setFileId(fileId);
+                } else {
+                    board.setFileId(0L);
+                }
+
+                boardRepository.save(board);
+
+                return new BoardRes(board);
             } else {
-                throw new BaseException(USERS_EMPTY_USER_ID);
+                throw new BaseException(POST_BOARD_FAILS);
             }
-        } catch (Exception exception) {
-            throw new BaseException(POST_BOARD_FAILS);
+
+        } catch (Exception exception){
+            throw new BaseException(DATABASE_ERROR);
         }
-
-
     }
 
     /* 게시글 수정 */
     @Transactional
-    public void updateBoard(long boardId, BoardReq update) throws BaseException {
+    public BoardRes updateBoard(Long boardId, BoardReq update) throws BaseException {
 
         try {
             Optional<Board> findBoard = boardRepository.findById(boardId);
-            System.out.println("--- update 전 ---");
-            System.out.println("title : " + findBoard.get().getTitle());
-            System.out.println("content : " + findBoard.get().getContent());
-            System.out.println("file id : " + findBoard.get().getFileId());
+            if (findBoard.isPresent()) {
+                Board board = findBoard.get();
 
-            findBoard.get().updateBoard(update);
-            System.out.println("--- update 후 ---");
-            System.out.println("title : " + findBoard.get().getTitle());
-            System.out.println("content : " + findBoard.get().getContent());
-            System.out.println("file id : " + findBoard.get().getFileId());
+                System.out.println("--- update 전 ---");
 
-            ResponseEntity.ok("-> 게시글 수정 완료");
+                if (update.getTitle() != null) { // 제목 변경
+                    System.out.println("title : " + board.getTitle());
+                    board.setTitle(update.getTitle());
+                }
+                if (update.getContent() != null) { // 내용 변경
+                    System.out.println("content : " + board.getContent());
+                    board.setContent(update.getContent());
+                }
+                if (update.getFileId() != null) { // 파일 변경
+                    System.out.println("file id : " + board.getFileId());
+                    board.setFileId(update.getFileId());
+                }
 
+                boardRepository.save(board);
+
+                return new BoardRes(board);
+            } else {
+                throw new BaseException(UPDATE_FAIL_BOARD);
+            }
         } catch (Exception exception){
             throw new BaseException(DATABASE_ERROR);
         }
@@ -224,18 +264,30 @@ public class BoardService {
             Optional<Board> findBoard = boardRepository.findById(boardId);
             if (findBoard.isPresent()) {
                 Long filedId = findBoard.get().getFileId();
-                fileRepository.deleteById(filedId);
+                if (filedId != 0) {
+                    fileRepository.deleteById(filedId);
+                } else {
+                    throw new BaseException(DELETE_FAIL_FILES);
+                }
                 boardRepository.deleteByBoardId(boardId);
                 ResponseEntity.ok("-> 게시글 삭제 완료");
             } else {
-                ResponseEntity.badRequest().body("not found : " + boardId);
+                throw new BaseException(BOARD_NOT_EXISTS);
             }
         } catch (Exception exception){
-            throw new BaseException(DATABASE_ERROR);
+            throw new BaseException(DELETE_FAIL_BOARD);
         }
-
-
     }
+
+    // ================================================================================
+
+//    public Page<Board> findAll(PageRequest pageRequest) {
+//        return boardRepository.findAll(pageRequest);
+//    }
+//
+//    public Page<Board> findByTitle(PageRequest pageRequest) {
+//        return boardRepository.findByTitle("1", pageRequest);
+//    }
 
 
 }
